게시물 전체 조회 함수 변경 / 게시글수정 & 게시글삭제 비밀번호 확인 함수 생성

+++ src/main/java/com/sparta/spring_w3_homework/controller/NoticeboardController.java
@@ -1,8 +1,13 @@
 package com.sparta.spring_w3_homework.controller;
 
+import java.sql.SQLOutput;
+import java.time.LocalDateTime;
 import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.sparta.spring_w3_homework.domain.Noticeboard;
+import com.sparta.spring_w3_homework.domain.NoticeboardRepository;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 import com.sparta.spring_w3_homework.domain.NoticeboardRequestDto;
@@ -11,10 +16,14 @@ import com.sparta.spring_w3_homework.service.NoticeboardService;
 
 import lombok.RequiredArgsConstructor;
 
+import static jdk.nashorn.internal.runtime.regexp.joni.Config.log;
+
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api")
 public class NoticeboardController {
+
+    private final NoticeboardRepository noticeboardRepository;
 
     private final NoticeboardService noticeboardService;
 
@@ -28,18 +37,17 @@ public class NoticeboardController {
 
     //게시글 전체 조회
     @GetMapping("/noticeboards")
-    public List<NoticeboardResponseDto> findAll(@RequestParam char deleteYn){
-        return noticeboardService.findAllByDeleteYn(deleteYn);
-    }
+    public List<Noticeboard> findAll(){
 
-    //게시글 수정
-    @PatchMapping("/noticeboards/{id}")
-    public Long update(@PathVariable Long id, @RequestBody NoticeboardRequestDto params){
+        //어제
+        LocalDateTime startDatetime = LocalDateTime.now().minusDays(1);
 
-//        noticeboardService.update(id, params);
-//        return id;
+        //오늘
+        LocalDateTime endDatetime = LocalDateTime.now();
 
-        return noticeboardService.update(id, params);
+        return noticeboardRepository.findAllByModifiedAtBetweenOrderByModifiedAtDesc(startDatetime, endDatetime);
+
+//        return noticeboardService.findAll();
     }
 
     //게시글 상세 조회
@@ -48,9 +56,27 @@ public class NoticeboardController {
         return noticeboardService.findById(id);
     }
 
+    //게시글 수정
+    @PatchMapping("/noticeboards/{id}")
+    public Long update(@PathVariable Long id, @RequestBody NoticeboardRequestDto params){
+        boolean result = noticeboardService.checkPw(id, params);
+        if(result){
+            noticeboardService.update(id, params);
+        } else {
+            System.out.println("비밀번호가 틀렸습니다.");
+        }
+        return id;
+    }
+
     //게시글 삭제
     @DeleteMapping("/noticeboards/{id}")
-    public Long delete(@PathVariable Long id){
-        return noticeboardService.delete(id);
+    public Long delete(@PathVariable Long id, @RequestBody NoticeboardRequestDto params){
+        boolean result = noticeboardService.checkPw(id, params);
+        if(result){
+            noticeboardService.delete(id);
+        } else {
+            System.out.println("비밀번호가 틀렸습니다.");
+        }
+        return id;
     }
 }
